+++ .github/workflows/integration-tests.yml
@@ -25,7 +25,7 @@ jobs:
         id: set-matrix
         run: |
           if [ x"${{ github.repository }}" == x"openai/triton" ]; then
-            echo '::set-output name=matrix::[["self-hosted", "A100"], ["self-hosted", "V100"], ["self-hosted", "gfx908"]]'
+            echo '::set-output name=matrix::[["self-hosted", "A100"], ["self-hosted", "V100"], ["self-hosted", "gfx908"], ["self-hosted", "arc770"]]'
           else
             echo '::set-output name=matrix::["ubuntu-latest"]'
           fi
@@ -53,6 +53,11 @@ jobs:
         run: |
           echo "BACKEND=ROCM" >> "${GITHUB_ENV}"
 
+      - name: Set XPU ENV
+        if: ${{(matrix.runner[0] == 'self-hosted') && (matrix.runner[1] == 'arc770')}}
+        run: |
+          echo "BACKEND=XPU" >> "${GITHUB_ENV}"
+
       - name: Clear cache
         run: |
           rm -rf ~/.triton
@@ -68,7 +73,7 @@ jobs:
           python3 -m pre_commit run --all-files
 
       - name: Install Triton
-        if: ${{ env.BACKEND != 'ROCM'}}
+        if: ${{ env.BACKEND == 'CUDA'}}
         run: |
           cd python
           python3 -m pip install --upgrade pip
@@ -84,8 +89,21 @@ jobs:
           python3 -m pip install torch==1.13.1 --index-url https://download.pytorch.org/whl/rocm5.2
           python3 -m pip install --no-build-isolation -vvv '.[tests]'
 
+      - name: Install Triton on XPU
+        if: ${{ env.BACKEND == 'XPU'}}
+        run: |
+          cd python
+          python3 -m pip install --upgrade pip
+          python3 -m pip install cmake==3.24
+          source ${HOME}/miniconda3/bin/activate
+          conda activate triton-xpu-ci
+          export TRITON_CODEGEN_INTEL_XPU_BACKEND=1
+          python3 -m pip uninstall -y triton
+          python3 setup.py build
+          python3 -m pip install --no-build-isolation -vvv '.[tests]'
+
       - name: Run lit tests
-        if: ${{ env.BACKEND != 'ROCM'}}
+        if: ${{ env.BACKEND == 'CUDA'}}
         run: |
           python3 -m pip install lit
           cd python
@@ -114,7 +132,7 @@ jobs:
           path: artifacts.tar.gz
 
       - name: Run CXX unittests
-        if: ${{ env.BACKEND != 'ROCM'}}
+        if: ${{ env.BACKEND == 'CUDA'}}
         run: |
           cd python
           cd "build/$(ls build | grep -i cmake)"
@@ -125,6 +143,15 @@ jobs:
         run: |
           cd python/test/unit/language
           python3 -m pytest --capture=tee-sys -rfs --verbose "test_core.py::test_empty_kernel"
+
+      - name: Run python tests on XPU
+        if: ${{ env.BACKEND == 'XPU'}}
+        run: |
+          source ${HOME}/triton_vars.sh
+          source ${HOME}/miniconda3/bin/activate
+          conda activate triton-xpu-ci
+          cd python/test/backend/third_party_backends
+          python3 -m pytest --capture=tee-sys -rfs --verbose --backend xpu
 
       - name: Regression tests
         if: ${{ contains(matrix.runner, 'A100') }}
